Defer to separate beatmap converter if 'crafted' tag found
